ayra2/requirements.txt
Flask==2.0.1
twilio==7.17.0
python-dotenv==0.19.2
groq==0.3.0
SpeechRecognition==3.8.1
gTTS==2.2.4
playsound==1.3.0
pandas==1.4.4
opencv-python==4.7.0.68
Pillow==9.3.0
google-generativeai==0.3.0
pyperclip==1.8.2
pyaudio==0.2.13
PyMySQL==1.1.0